+++ libs/adv_logger_ntp/adv_logger_ntp.h
@@ -1,7 +1,7 @@
 /*
  * Advanced ESP Logger with NTP support
  *
- * Copyright 2020-2021 José Antonio Jiménez Campos (@RavenSystem)
+ * Copyright 2020-2022 José Antonio Jiménez Campos (@RavenSystem)
  *
  */
 
@@ -32,7 +32,7 @@ extern "C" {
 #define ADV_LOGGER_UART1_UDP_BUFFERED   (8)
 
 
-void adv_logger_init(const uint8_t log_type, char* dest_addr);
+void adv_logger_init(const int log_type, char* dest_addr);
 
 int adv_logger_remove();
