+++ libs/adv_nrzled/adv_nrzled.c
@@ -1,7 +1,7 @@
 /*
 * Advanced NRZ LED Driver
 *
-* Copyright 2021 José Antonio Jiménez Campos (@RavenSystem)
+* Copyright 2021-2022 José Antonio Jiménez Campos (@RavenSystem)
 *
 */
 
@@ -12,8 +12,8 @@
 #define CPU_FREQ_MHZ                        (80)
 #define NRZ_TICKS_OFFSET                    (4)
 
-uint16_t nrz_ticks(const float time_us) {
-    uint16_t cycles = CPU_FREQ_MHZ * time_us;
+int nrz_ticks(const float time_us) {
+    int cycles = CPU_FREQ_MHZ * time_us;
     
     if (cycles < NRZ_TICKS_OFFSET) {
         cycles = NRZ_TICKS_OFFSET;
@@ -32,10 +32,10 @@ void IRAM nrzled_set(const uint8_t gpio, const uint16_t ticks_0, const uint16_t 
     uint32_t c, t;
     uint32_t start = 0;
     
-    for (uint16_t i = 0; i < size; i++) {
+    for (int i = 0; i < size; i++) {
         uint8_t color = *colors++;
         
-        for (int8_t p = 7; p >= 0; p--) {
+        for (int p = 7; p >= 0; p--) {
             t = ticks_0;
             if (color & (1 << p)) {
                 t = ticks_1;
