+++ libs/adv_pwm/adv_pwm.h
@@ -1,7 +1,7 @@
 /*
 * Advanced PWM Driver
 *
-* Copyright 2021 José Antonio Jiménez Campos (@RavenSystem)
+* Copyright 2021-2022 José Antonio Jiménez Campos (@RavenSystem)
 *
 */
 
@@ -14,11 +14,11 @@ extern "C" {
 
 void adv_pwm_start();
 void adv_pwm_stop();
-void adv_pwm_set_freq(const uint16_t freq);
-void adv_pwm_set_duty(const uint8_t gpio, const uint16_t duty, uint16_t dithering);
-uint16_t adv_pwm_get_duty(const uint8_t gpio);
-void adv_pwm_new_channel(const uint8_t gpio, const bool inverted);
-//void adv_pwm_set_zc_gpio(const uint8_t gpio);
+void adv_pwm_set_freq(const unsigned int freq);
+void adv_pwm_set_duty(const unsigned int gpio, const unsigned int duty, unsigned int dithering);
+int adv_pwm_get_duty(const unsigned int gpio);
+void adv_pwm_new_channel(const unsigned int gpio, const bool inverted);
+//void adv_pwm_set_zc_gpio(const unsigned int gpio);
 
 #ifdef __cplusplus
 }
