+++ libs/raven_ntp/raven_ntp.h
@@ -1,9 +1,9 @@
 /*
-* Advanced HLW8012 Driver
-*
-* Copyright 2020 José Antonio Jiménez Campos (@RavenSystem)
-*
-*/
+ * RavenSystem NTP
+ *
+ * Copyright 2021-2022 José Antonio Jiménez Campos (@RavenSystem)
+ *
+ */
 
 #ifndef __RAVEN_NTP_H__
 #define __RAVEN_NTP_H__
@@ -16,7 +16,7 @@ extern "C" {
 
 int raven_ntp_update(char* ntp_server);
 time_t raven_ntp_get_time_t();
-void raven_ntp_get_log_time(char* buffer, const size_t buffer_size);
+void raven_ntp_get_log_time(char* buffer, const unsigned int buffer_size);
 
 #ifdef __cplusplus
 }
